+++ scalapb-runtime/non-jvm/src/main/scala/com/google/protobuf/field_mask/FieldMaskProto.scala
@@ -13,10 +13,10 @@ object FieldMaskProto extends _root_.scalapb.GeneratedFileObject {
   )
   private lazy val ProtoBytes: Array[Byte] =
       scalapb.Encoding.fromBase64(scala.collection.Seq(
-  """CiBnb29nbGUvcHJvdG9idWYvZmllbGRfbWFzay5wcm90bxIPZ29vZ2xlLnByb3RvYnVmIiEKCUZpZWxkTWFzaxIUCgVwYXRoc
-  xgBIAMoCVIFcGF0aHNCiQEKE2NvbS5nb29nbGUucHJvdG9idWZCDkZpZWxkTWFza1Byb3RvUAFaOWdvb2dsZS5nb2xhbmcub3JnL
-  2dlbnByb3RvL3Byb3RvYnVmL2ZpZWxkX21hc2s7ZmllbGRfbWFza6ICA0dQQqoCHkdvb2dsZS5Qcm90b2J1Zi5XZWxsS25vd25Ue
-  XBlc2IGcHJvdG8z"""
+  """CiBnb29nbGUvcHJvdG9idWYvZmllbGRfbWFzay5wcm90bxIPZ29vZ2xlLnByb3RvYnVmIi0KCUZpZWxkTWFzaxIgCgVwYXRoc
+  xgBIAMoCUIK4j8HEgVwYXRoc1IFcGF0aHNCiQEKE2NvbS5nb29nbGUucHJvdG9idWZCDkZpZWxkTWFza1Byb3RvUAFaOWdvb2dsZ
+  S5nb2xhbmcub3JnL2dlbnByb3RvL3Byb3RvYnVmL2ZpZWxkX21hc2s7ZmllbGRfbWFza6ICA0dQQqoCHkdvb2dsZS5Qcm90b2J1Z
+  i5XZWxsS25vd25UeXBlc2IGcHJvdG8z"""
       ).mkString)
   lazy val scalaDescriptor: _root_.scalapb.descriptors.FileDescriptor = {
     val scalaProto = com.google.protobuf.descriptor.FileDescriptorProto.parseFrom(ProtoBytes)