+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 
-## [0.8.3](https://github.com/scalapb/ScalaPB/tree/HEAD)
+## [0.8.4](https://github.com/scalapb/ScalaPB/tree/HEAD)
+- Add `scalaName` to FieldDescriptor to make the name of the Scala field
+  accessible from the descriptor. This is done by updating
+  `scalapb.FieldOptions` in the descriptor we serialize in the proto file
+  companion object.
+
+## [0.8.3](https://github.com/scalapb/ScalaPB/tree/v0.8.3)
+[Full Changelog](https://github.com/scalapb/ScalaPB/compare/v0.8.2...v0.8.3)
 - Add `retain_source_code_info` setting for accessing source code information through
   the `location` accessor on descriptors.
 - Add option to not generate lenses.